IR-224: Use `this`` consistently in Mapping.kt
The extension functions on the `Report`` JPA entity were
sometimes using `this` (the report), sometimes not
because it's implicit.

I've updated them so that they all use `this` explicitly.
`this` is not the most beautiful thing and code is nicer
when it's omitted. However using explicit `this` adds to
clarity and avoid any confusion of who's the subject.