IR-224: NOMIS Sync: Update even more fields
Some of the fields in the Report were not updated yet.
If the NOMIS sync request explicitly contains a value
we should honour this and update the Report.

A good example of this is the `reportingStaff` and
`reportedDateTime` fields in the sync request.

In other cases we try to do something as similar as
possible/sensible to the the logic in `Mapping.kt`.

A good example of this is the `questionSetId` field
in the report: This is the `questionnaireId` in the sync request
converted to string. There is an implicit relationship between
the two but we still update the field because when this value
change the whole set of questions change and it would
retain the existing `questionSetId` wouldn't make much sense.
(This is a new field and it may not really matter much)

The `Event` is also a grey area because it's a new concept not
present in NOMIS (and currently not used anyway).
As such the NOMIS sync request doesn't contain any information
on how to update the event.

Normally we wouldn't want to implicitly update the Event record
but we have decided that the pragmatic thing to do, at least for now,
is to update the events field when we update the corresponding
record event.
For example, title/description of the `Report` are updated,
the `Event`'s title/description are generated from the `Report`
when it's created so on update we also update these.

In practice it may not really matter on migration from NOMIS
each Report will just have its own Event anyway. And furthermore
a given Report would only be updated either in DPS or in old NOMIS.
Even in the unlikely/rare case where a Report may be updated both
in DPS and NOMIS the risk is still low. If we find that inferring
the new values of the `Event` fields from `Report` fields is problematic
we can always break this link in the future.