IR-224: NOMIS sync endpoint updates report (#80)
* Fix test data in SyncServiceTest where `sampleReport` was incorrectly listing staff involvement as specified in `sampleSyncRequest`

* IR-224: Upsert endpoint updates type

* IR-224: Upsert endpoint updates staff involvment

* IR-224: Fix SyncServiceTest prisoner involvment test data

* IR-224: Upsert endpoint updates prisoner involvment

* Refactored NOMIS Mapping logic to be reused when upserting

Upsert/updateWith logic is basically the same as to when converting a
`NomisReport` DTO class to a `Report` JPA entity (but without the clearing)

For this reason I've moved this logic in building blocks that can be reused
by the `updateWith()` method.

* IR-224: Upsert endpoint updates correction requests

NOTE: These are called "requirements" in the NOMIS world for some reason.

* Fixed SyncService test data and test

* IR-224: Upsert endpoint updates questions/responses

* Fixed SyncServiceTest's questions/answers data

* IR-224: NOMIS sync endpoint updates history

history is updated using the `history` field provided in the sync request.

* IR-224: Removed `equals()` from `StaffInvolvement`

This custom `equals()` implementation caused problems
with the upsert. When sync request contained multiple
`StaffInvolvement` objects only one was persisted.

After removing the custom `equals()` JPA started to
do what we expect it to do.

* IR-224: Removed `equals()` from `PrisonerInvolvement`

Custom implementation of `equals()` method in the `PrisonerInvolvement`
JPA entity class caused JPA persistence to not work.

When adding multiple prisoner involvements to a report
only one was actually persisted.

* IR-224: Removed `equals()` from `CorrectionRequest`

This was causing persistence problems. JPA would
only persist one of the correction requests.

* IR-224: Removed harmful `equals()` from Question/Response

Questions/Responses were not persisted correctly by JPA
because of the bogus implementation of `equals()`.

* IR-224: Removed harmful `equals()` from History/HistoricalQuestion/HistoricalResponse

History was not persisted correctly by JPA because of the bogus
implementation of `equals()`.

* IR-224: Removed remaining custom `equals()` implementations

These don't cause any tests to fail (yet) but that's
probably because we don't have much logic for these new
concepts yet.

* IR-224: Use `this`` consistently in Mapping.kt

The extension functions on the `Report`` JPA entity were
sometimes using `this` (the report), sometimes not
because it's implicit.

I've updated them so that they all use `this` explicitly.
`this` is not the most beautiful thing and code is nicer
when it's omitted. However using explicit `this` adds to
clarity and avoid any confusion of who's the subject.

* IR-224: NOMIS Sync: Update even more fields

Some of the fields in the Report were not updated yet.
If the NOMIS sync request explicitly contains a value
we should honour this and update the Report.

A good example of this is the `reportingStaff` and
`reportedDateTime` fields in the sync request.

In other cases we try to do something as similar as
possible/sensible to the the logic in `Mapping.kt`.

A good example of this is the `questionSetId` field
in the report: This is the `questionnaireId` in the sync request
converted to string. There is an implicit relationship between
the two but we still update the field because when this value
change the whole set of questions change and it would
retain the existing `questionSetId` wouldn't make much sense.
(This is a new field and it may not really matter much)

The `Event` is also a grey area because it's a new concept not
present in NOMIS (and currently not used anyway).
As such the NOMIS sync request doesn't contain any information
on how to update the event.

Normally we wouldn't want to implicitly update the Event record
but we have decided that the pragmatic thing to do, at least for now,
is to update the events field when we update the corresponding
record event.
For example, title/description of the `Report` are updated,
the `Event`'s title/description are generated from the `Report`
when it's created so on update we also update these.

In practice it may not really matter on migration from NOMIS
each Report will just have its own Event anyway. And furthermore
a given Report would only be updated either in DPS or in old NOMIS.
Even in the unlikely/rare case where a Report may be updated both
in DPS and NOMIS the risk is still low. If we find that inferring
the new values of the `Event` fields from `Report` fields is problematic
we can always break this link in the future.

* IR-224: Consolidated some of the TODO comments for the event

---------

Co-authored-by: Igor <igor.ushkarev@digital.justice.gov.uk>