IR-245: DTOs: Renamed fields for consistency
In the non-NOMIS DTO classes renamed metadata fields to be more consistent:
- `createdDate` => `createdAt` (it's a datetime)
- `lastModifiedDate` => `modifiedAt` (last is implicit)
- `lastModifiedBy` => `modifiedBy` (last is implicit)
- `changeDate` => `changedAt` (it's a datetime)
- `changeStaffUsername` => `changedBy` (staff is implicit)
- `reportedDate` => `reportedAt` (it's a datetime)
- `recordedOn` => `recordedAt`
- `setOn`/`setBy` => `changedAt`/`changedBy` (`StatusHistory`
   consistent with `History`)

NOTE: This is a **breaking change** as the API responses change.
NOTE: This is not touching the JPA Entities just yet,
but that will require changing the DB schema. It's probably worth doing
at this stage for consistency.