IR-256: Listen to (but ignore) 'prisoner.merged' events
We are currently subscribed to these domain events but we didn't do
anything with them.
This means messages accumulate in the queue and we're getting some
alerts.

This should stop these alerts from happening.

PS: We will eventually have to introduce some logic to handle the merge
of two prisoner numbers (e.g. updating prisoner involvement) but that's
for another day.