IRS-291: Add `whatChanged` to additional info
Added optional `whatChanged` field to the domain events'
`additionalInformation` field.

The idea is that the consumer will be able to have a
better idea of what part of the report changed.

Most of these types shouldn't need an explaination
but there are a few "special" ones:
- `BASIC_REPORT` this is for when only the basic fields
  of a report were changed (e.g. not the associated records
  in other DB table)
- `ANYTHING` will be mainly used for when the sync
  endpoint will be used and we get the big report blob.
  This means anything, and potentially multiple things
  changed, both in the basic report fields and in any
  of the associated records (e.g. Prisoners involved
  or what not)

NOTE: This enum may need tweaking, for example we currently
don't have an endpoint to update questions/answers. I've
added a `QUESTIONS` discriminant but we may want to rename
this or split into `QUESTIONS`` and `ANSWERS` depending
on what we're going to do with the API.