IRS-291: Upsert endpoint populate `whatChanged` field
When the sync upsert endpoint is used to update an existing Report the
published domain event has a `whatChanged` field populated.

Because this endpoint is very broad and the client pass the whole Report
blob and could change any part of the report we don't have any trivial
way of knowing which specific part was updated.
For this reason the `whatChanged` field is simply set to `ANYTHING` as
any part of the report may have changed.

It may not be super useful but at least it's better than just being `null`
that is even more ambiguous.