Removed locations endpoints
The location in the service will not be as structured as this and values
are almost free-text. This "location" concept in the API is new and was
not there in NOMIS and we think we may not need it.
So removing this now, when it's easy as we don't have any consumers.