IR-577: Removed `QID` prefix from question codes
This is not necessary and it may well be in our way when updating answers.
Regardless it makes the UI a bit more fiddly because we have to map back
and forth between question IDs in the configs and questions code with the
prefix (which is not guaranteed to be there for future report types config).

It is now a good time to remove this as data it not impacting real users
(only clients are syncing/reconciliation ones)