Try and retain existing description addendums records
Refactored the logic that adds description addendums
to a report to try and retain existing records whenever
possible.

This is similar to what we do for other related objects
where equality is used to determine if there is already
a record for the object in the sync request.

This means on update the records are not always
cleared/removed and then readded. It is particularly
interesting in this case because description addendums
are always added and not updated/removed (although
the logic honour the upsert request and could
potentially remove some addendums if the sync request
doesn't have them anymore)