initial function creation for description addendum extraction (#315)
* initial function creation for description addendum extraction

* Tweaks to get existing tests passing

- removed `NomisReport` dto class
- combined `getBaseDescription()`/`getAddendums()`
  methods
- reuse existing method to add description addendum

NOTE: Kotlin complains if you have a method called
`getAddendums` because it thinks it's a getter for
a property that doesn't exist.

* IR-1001: Endpoint to add a description addendum (#314)

* IR-1001: Endpoint to add a description addendum

* Get report "eagerly" when adding description addendums

The report will be converted to `ReportWithDetails`
so it will need the related objects anyway so it should
be more efficient to get these in less queries.

* Added new test for splitting NOMIS descriptions

* Added/changed more unit tests and renamed functions following change.

* IR-1000 Refactor incident description formatting in NomisSyncResourceTest

* Refactor date-time parsing in NomisReport to improve readability and maintainability

* Refactor date-time parsing in NomisReport to improve readability and maintainability

* Added validation exceptions to getDecriptionParts function and created unit tests

* Linting changes

* Improved description addendum sync tests

Note: NOMIS description datetimes granularity means
time doesn't go beyond seconds, so test description
addendums also have time truncated to the seconds
to be sure we can compare datetimes.

* Try and retain existing description addendums records

Refactored the logic that adds description addendums
to a report to try and retain existing records whenever
possible.

This is similar to what we do for other related objects
where equality is used to determine if there is already
a record for the object in the sync request.

This means on update the records are not always
cleared/removed and then readded. It is particularly
interesting in this case because description addendums
are always added and not updated/removed (although
the logic honour the upsert request and could
potentially remove some addendums if the sync request
doesn't have them anymore)

---------

Co-authored-by: Aldo "xoen" Giambelluca <aldo.giambelluca@digital.justice.gov.uk>
Co-authored-by: Aldo Giambelluca <xoen@users.noreply.github.com>
Co-authored-by: Michael Willis <michael.willis@digital.justice.gov.uk>