Increased memory across all the environments
Since yesterday afternoon we're seeing a lot of errors
in `dev` and pods restarts.

Pods logs have a bunch of errors like:
`java.lang.OutOfMemoryError: Java heap space`

While the RDS instance resource usage seems to be
fine, but the logs also have a bunch of errors along
the line of "connection reset by peer", again, pointing
in the direction of the API pods rather than a problem
with the DB sizing (AFAICT).

Production memory was already [tweaked in March] possibly
because of a similar problem, so I'm updating the
memory across all the environments, specifically the
Java maximum heap memory allocation.

[tweaked in March]: https://github.com/ministryofjustice/hmpps-incident-reporting-api/commit/d059c3e8126d38661e1c21d241142b1750f54673