Revert "Wait 3 days before using a new version of a dependency"
Because of the rolling nature of these updates, waiting 3 days for a
new version of a package potentially can cause a situation where we
get the broken version but not the (fixed) subsequent one as it's too
fresh.

So maybe we just use our judgment instead of having an hard rule in the
Renovate config.

This reverts commit 3aedb2f813a0958be57651484ef30e2bf1659be7.