Renovate tweaks (#380)
* Wait 3 days before using a new version of a dependency

This limits the chance of using a bogus new release.

(unless it's a security fix)

* Bump swagger-parser separately

New version of this package have been breaking our tests
so I'm instructing renovate to bump this separately so that we don't
have to manually tweak the PR that updates all non-major dependencies.

* Bump hmpps-digital-prison-reporting-lib separately

This library has been a bit temperamental in the past and sometimes
even non-major version changes have introduced breaking changes.

I've instructing Renovate to bump this separately.

* Revert "Wait 3 days before using a new version of a dependency"

Because of the rolling nature of these updates, waiting 3 days for a
new version of a package potentially can cause a situation where we
get the broken version but not the (fixed) subsequent one as it's too
fresh.

So maybe we just use our judgment instead of having an hard rule in the
Renovate config.

This reverts commit 3aedb2f813a0958be57651484ef30e2bf1659be7.